+++ internal/confmapprovider/discovery/bundle/bundledfs_windows.go
@@ -1,4 +1,4 @@
-// Copyright  Splunk, Inc.
+// Copyright Splunk, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -21,12 +21,10 @@ import (
 )
 
 // BundledFS is the in-executable filesystem that contains all bundled discovery config.d components.
-//
-// If you are bootstrapping bundle_gen.go or the `discoverybundler` cmd without any rendered files in bundle.d,
-// comment out the below embed directives before installing to prevent "no matching files found"
-// build errors.
-//
-//go:embed bundle.d/extensions/*.discovery.yaml
+
+//go:embed bundle.d/extensions/docker-observer.discovery.yaml
+//go:embed bundle.d/extensions/host-observer.discovery.yaml
+//go:embed bundle.d/extensions/k8s-observer.discovery.yaml
 //go:embed bundle.d/receivers/oracledb.discovery.yaml
 //go:embed bundle.d/receivers/postgresql.discovery.yaml
 //go:embed bundle.d/receivers/redis.discovery.yaml