+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - (Splunk) Update splunk-otel-js to [v2.6.0](https://github.com/signalfx/splunk-otel-js/releases/tag/v2.6.0) ([#4064](https://github.com/signalfx/splunk-otel-collector/pull/4064))
 
+### 🛑 Breaking changes 🛑
+
+- `postgresql` discovery now uses the OpenTelemetry Collector Contrib receiver by default instead of the smartagent receiver.
+
 ## v0.91.0
 
 This Splunk OpenTelemetry Collector release includes changes from the [opentelemetry-collector v0.91.0](https://github.com/open-telemetry/opentelemetry-collector/releases/tag/v0.91.0) and the [opentelemetry-collector-contrib v0.91.0](https://github.com/open-telemetry/opentelemetry-collector-contrib/releases/tag/v0.91.0) releases where appropriate.